Fix globalThis so it is populated before imports and therefore can be used immediately anywhere
