+++ src/renderer/globalThis.ts
@@ -1,0 +1,7 @@
+/**
+ * Module to set up globalThis and polyfills in the renderer
+ */
+
+import { ProcessType } from '@shared/globalThis';
+
+globalThis.processType = ProcessType.Renderer;
